+++ src/main/java/com/atlauncher/data/minecraft/loaders/neoforge/NeoForgeLoader.java
@@ -83,8 +83,7 @@ public class NeoForgeLoader implements Loader {
         String artifactName = is1201Version ? "forge" : "neoforge";
         String versionName = is1201Version ? this.minecraft + "-" + this.version : this.rawVersion;
 
-        this.installerPath = FileSystem.LOADERS
-                .resolve("forge-" + this.minecraft + "-" + this.version + "-installer.jar");
+        this.installerPath = FileSystem.LOADERS.resolve(artifactName + "-" + versionName + "-installer.jar");
         this.installerUrl = Constants.NEOFORGE_MAVEN + "/net/neoforged/" + artifactName + "/" + versionName + "/"
                 + artifactName + "-" + versionName + "-installer.jar";
     }