+++ CHANGELOG.md
@@ -10,5 +10,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 ### Fixes
 - NPE by NewsTab.reload
 - Issues installing Forge 1.20.4
+- Launch process trying to download or continue with libraries that don't exist and cannot be downloaded
 
 ### Misc