+++ CHANGELOG.md
@@ -11,5 +11,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - NPE by NewsTab.reload
 - Issues installing Forge 1.20.4
 - Launch process trying to download or continue with libraries that don't exist and cannot be downloaded
+- NeoForge installer files not being saved with correct name
 
 ### Misc